SONAR-10040 move projectbranch ws to package branch
