+++ server/sonar-server/src/main/java/org/sonar/server/branch/ws/BranchWsModule.java
@@ -17,7 +17,17 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-@ParametersAreNonnullByDefault
-package org.sonar.server.projectbranch.ws;
+package org.sonar.server.branch.ws;
 
-import javax.annotation.ParametersAreNonnullByDefault;
+import org.sonar.core.platform.Module;
+
+public class BranchWsModule extends Module {
+  @Override
+  protected void configureModule() {
+    add(
+      ListAction.class,
+      DeleteAction.class,
+      RenameAction.class,
+      BranchesWs.class);
+  }
+}
