+++ server/sonar-server/src/test/java/org/sonar/server/branch/ws/BranchesWsTest.java
@@ -17,19 +17,35 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.projectbranch.ws;
+package org.sonar.server.branch.ws;
 
 import org.junit.Test;
-import org.sonar.core.platform.ComponentContainer;
+import org.sonar.api.server.ws.Request;
+import org.sonar.api.server.ws.Response;
+import org.sonar.api.server.ws.WebService;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.sonar.core.platform.ComponentContainer.COMPONENTS_IN_EMPTY_COMPONENT_CONTAINER;
 
-public class BranchWsModuleTest {
+public class BranchesWsTest {
+
   @Test
-  public void verify_count_of_added_components() {
-    ComponentContainer container = new ComponentContainer();
-    new BranchWsModule().configure(container);
-    assertThat(container.size()).isEqualTo(COMPONENTS_IN_EMPTY_COMPONENT_CONTAINER + 4);
+  public void define_ws() {
+    BranchesWs underTest = new BranchesWs(new BranchWsAction() {
+      @Override
+      public void define(WebService.NewController context) {
+        context.createAction("foo").setHandler(this);
+      }
+
+      @Override
+      public void handle(Request request, Response response) throws Exception {
+
+      }
+    });
+
+    WebService.Context context = new WebService.Context();
+    underTest.define(context);
+
+    assertThat(context.controller("api/project_branches").action("foo")).isNotNull();
   }
+
 }
