+++ server/sonar-server/src/test/java/org/sonar/server/branch/ws/DeleteActionTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.projectbranch.ws;
+package org.sonar.server.branch.ws;
 
 import org.junit.Rule;
 import org.junit.Test;
